Remove `uses sled` from readme (#106)
The readme shows some dates information, namely that `sled` is used for storage. Most people won't really care for this, but for me this was a particular advantage as I'd like to embed this library into a project that already uses sled.

I haven't removed `sled` from line 240 - I don't know how that should change.